fix: surface real MinIO stat faults and map bad path/sort params to 400

+++ src/main/java/com/clara/ops/challenge/document_management_service_challenge/web/dto/PaginationMetadata.java
@@ -1,4 +1,4 @@
 package com.clara.ops.challenge.document_management_service_challenge.web.dto;
 
 public record PaginationMetadata(
-    int currentPage, int itemsPerPage, int currentItems, int totalPages, long totalItems) {}
+    int currentPage, int itemsPerPage, int currentItems, int totalPages, int totalItems) {}
